# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: docz-api
  version: 1.2.1
  description: |
    HTTP contract for docz-api's `/api/v1` read and search surface, plus the
    site auth endpoints and the GitHub webhook receiver. Hand-authored and kept
    honest by an in-process kin-openapi contract test (see IMPL-0002).
servers:
  - url: /
    description: Same-origin. The docz-site and the API share an origin.
...
security:
  - sessionCookie: []
paths:
...
  /api/v1/repos/{owner}/{name}/types/{type}/docs:
    parameters:
      - $ref: "#/components/parameters/Owner"
      - $ref: "#/components/parameters/Name"
      - $ref: "#/components/parameters/Type"
    get:
      tags: [docs]
      operationId: listDocs
      summary: List documents of a type (without raw_md).
      responses:
        "200":
          description: Documents of the given type. Each omits `raw_md`.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [docs]
                properties:
                  docs:
                    type: array
                    items:
                      $ref: "#/components/schemas/Document"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/repos/{owner}/{name}/types/{type}/docs/{doc_id}:
    parameters:
      - $ref: "#/components/parameters/Owner"
      - $ref: "#/components/parameters/Name"
      - $ref: "#/components/parameters/Type"
      - $ref: "#/components/parameters/DocID"
    get:
      tags: [docs]
      operationId: getDoc
      summary: Fetch a single document, including raw_md.
      responses:
        "200":
          description: The document, including `raw_md`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /webhooks/github:
    post:
      tags: [webhooks]
      operationId: githubWebhook
      summary: Receive a GitHub App webhook delivery.
      description: |
        Authenticated by an HMAC-SHA256 signature over the raw request body (the
        `X-Hub-Signature-256` header), verified in constant time — there is no
        session or bearer auth. Verified deliveries are deduplicated by
        `X-GitHub-Delivery`; the payload is GitHub's own event schema (validated
        by go-github at runtime, opaque here) and processing is asynchronous, so
        a valid delivery returns 202 immediately.
      security: []
      parameters:
        - name: X-Hub-Signature-256
          in: header
          required: true
          description: Hex-encoded "sha256=…" HMAC of the raw body under the app secret.
          schema:
            type: string
        - name: X-GitHub-Event
          in: header
          required: true
          description: The event type (installation, push, release, …).
          schema:
            type: string
        - name: X-GitHub-Delivery
          in: header
          required: true
          description: Unique delivery id, used for idempotent de-duplication.
          schema:
            type: string
      requestBody:
        required: true
        description: The GitHub event payload (opaque; validated by go-github).
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "202":
          description: Delivery accepted for asynchronous processing.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "400":
          description: Unreadable body or unparseable payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          description: Missing or invalid HMAC signature.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: docz_session
      description: |
        Opaque Redis-backed session id set by the auth flow. Sent as the
        `docz_session` cookie; required on every `/api/v1` operation unless the
        operation overrides `security` (public routes use `security: []`).

        A server started with `AUTH_PROVIDERS=none` (the login-free first-setup
        mode) accepts every request without a cookie and serves it as an
        anonymous identity, and does not mount `/auth/login` or
        `/auth/callback`. Request and response shapes are unchanged, so a client
        generated from this spec works against either mode.
  parameters:
    Owner:
      name: owner
      in: path
      required: true
      description: Repository owner (org or user login).
      schema:
        type: string
      example: acme
    Name:
      name: name
      in: path
      required: true
      description: Repository name.
      schema:
        type: string
      example: platform
    Type:
      name: type
      in: path
      required: true
      description: Doc type, addressable by canonical name, id_prefix, or alias.
      schema:
        type: string
      example: frameworks
    DocID:
      name: doc_id
      in: path
      required: true
      description: Document id.
      schema:
        type: string
      example: FW-0001
  responses:
    Unauthorized:
      description: Missing or invalid session.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: |
        Repository, type, document, or repository index not found. Also
        returned for repositories outside the caller's allowed set, to hide
        their existence.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      description: A JSON error envelope with a single human-readable message.
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
          example: repo not found
...
    Document:
      description: A single document. `raw_md` is present only on the get-one endpoint.
      type: object
      additionalProperties: false
      required: [repo, doc_id, type, title, status, author, created, path, git_sha, content_hash, updated_at]
      properties:
        repo:
          type: string
          example: acme/platform
        doc_id:
          type: string
          example: FW-0001
        type:
          type: string
          description: Canonical type name.
          example: frameworks
        title:
          type: string
        status:
          type: string
          description: '"" when unset.'
        author:
          type: string
          description: '"" when unset.'
        created:
          type: string
          description: 'YYYY-MM-DD, or "" when unset.'
        path:
          type: string
          example: docs/frameworks/0001-intro.md
        git_sha:
          type: string
        content_hash:
          type: string
        updated_at:
          type: string
          description: 'RFC3339, or "" when unset.'
        raw_md:
          type: string
          description: Full Markdown source; present only on the single-document endpoint.
...
    StatusResponse:
      description: A minimal JSON status envelope with a single `status` field.
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          description: A short human-readable status word or phrase.
          example: logged out