# cmd/docz-api/main.go
// Command docz-api is the entry point for the docz-api service.
//
// It stays thin: parse flags, load configuration, configure the slog default
// handler, wire dependencies, and run an HTTP server with graceful shutdown.
// All real behavior lives under internal/.
package main

import (
	"context"
	"encoding/json"
	"errors"
	"flag"
	"fmt"
	"log/slog"
	"net/http"
	"os"
	"os/signal"
	"strconv"
	"strings"
	"syscall"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"

	"github.com/donaldgifford/docz-api/api"
	"github.com/donaldgifford/docz-api/internal/auth"
	"github.com/donaldgifford/docz-api/internal/authhttp"
	"github.com/donaldgifford/docz-api/internal/authorize"
	"github.com/donaldgifford/docz-api/internal/config"
	"github.com/donaldgifford/docz-api/internal/githubapp"
	"github.com/donaldgifford/docz-api/internal/httpapi"
	"github.com/donaldgifford/docz-api/internal/queue"
	"github.com/donaldgifford/docz-api/internal/search"
	"github.com/donaldgifford/docz-api/internal/session"
	"github.com/donaldgifford/docz-api/internal/store"
	"github.com/donaldgifford/docz-api/internal/telemetry"
	"github.com/donaldgifford/docz-api/internal/webhook"
)

// Build metadata, injected via -ldflags at release time (see justfile).
var (
	version = "dev"
	commit  = "none"
	date    = "unknown"
)

// Server timeouts. readHeaderTimeout bounds slow-header (Slowloris) clients;
// shutdownTimeout bounds in-flight request draining on shutdown; readyzTimeout
// bounds the dependency check behind the readiness probe.
const (
	readHeaderTimeout = 10 * time.Second
	shutdownTimeout   = 15 * time.Second
	readyzTimeout     = 2 * time.Second
)

// namedChecker pairs a downstream dependency's name with its readiness check.
// The readiness probe runs each one and reports per-dependency status, so an
// outage names the offender.
type namedChecker struct {
	name  string
	check func(ctx context.Context) error
}

func main() {
	if err := run(); err != nil {
		// run() configures slog once config loads; a failure before that still
		// reaches the bootstrap default handler.
		slog.Error("docz-api exited with error", "err", err)
		os.Exit(1)
	}
}

// run parses flags, loads config, configures logging, and serves until a
// termination signal arrives.
func run() error {
	showVersion := flag.Bool("version", false, "print version information and exit")
	migrateOnly := flag.Bool("migrate", false, "apply database migrations and exit")
	onboardSpec := flag.String("onboard", "",
		"seed a repo and run one ingest, then exit: owner/name@installation_id")
	flag.Parse()

	if *showVersion {
		fmt.Printf("docz-api %s (commit %s, built %s)\n", version, commit, date)
		return nil
	}

	cfg, err := config.Load()
	if err != nil {
		return fmt.Errorf("loading config: %w", err)
	}

	logger, err := newLogger(cfg.Log)
	if err != nil {
		return err
	}
	slog.SetDefault(logger)

	// Telemetry (OQ 8): install the global tracer + W3C propagator and register
	// the Prometheus instruments. Non-blocking (the OTLP/HTTP exporter connects
	// lazily); a no-op when no collector endpoint is configured.
	telShutdown, err := telemetry.Setup(context.Background(), telemetry.Config{
		ServiceName:    cfg.Telemetry.ServiceName,
		OTLPEndpoint:   cfg.Telemetry.OTLPEndpoint,
		SampleRate:     cfg.Telemetry.SampleRate,
		MetricsEnabled: cfg.Telemetry.MetricsEnabled,
	})
	if err != nil {
		return fmt.Errorf("setting up telemetry: %w", err)
	}
	defer shutdownTelemetry(telShutdown)

	// Apply pending migrations before serving so the schema is always current.
	// `-migrate` runs them and exits — a CI/ops pre-deploy step.
	if err := store.Migrate(context.Background(), cfg.Store.DatabaseURL); err != nil {
		return fmt.Errorf("applying migrations: %w", err)
	}
	if *migrateOnly {
		slog.Info("migrations applied; exiting")
		return nil
	}

	// Prove the App credentials before anything depends on them. They are
	// otherwise first exercised inside a worker job, where a bad key looked like
	// an unexplained ingest failure (INV-0007 F5). Skipped for -migrate, which
	// never talks to GitHub.
	if err := checkGitHubCredentials(context.Background(), cfg.GitHub); err != nil {
		return err
	}

	// Runtime connection pool (separate from the migration connection). Owned
	// here for the process lifetime; closed on shutdown.
	pool, err := store.NewPool(context.Background(), cfg.Store.DatabaseURL)
	if err != nil {
		return fmt.Errorf("connecting to postgres: %w", err)
	}
	defer pool.Close()
	st := store.NewStore(pool)

	// Search client: ensure the documents index + settings exist before any
	// ingest writes to it or the search endpoint reads from it.
	searchClient := search.New(cfg.Meili.Host, cfg.Meili.APIKey.Reveal())
	if err := searchClient.EnsureIndex(context.Background()); err != nil {
		return fmt.Errorf("ensuring search index: %w", err)
	}

	// Async ingest queue: the enqueue client is shared; the worker (started
	// below for the server path) drains jobs through the ingest pipeline.
	queueClient, err := queue.NewClient(cfg.Store.RedisURL, cfg.Ingest.Debounce)
	if err != nil {
		return fmt.Errorf("connecting to redis: %w", err)
	}

	// `-onboard` seeds the installation and enqueues one ingest, then exits — the
	// manual trigger now goes through the queue (a running server drains it).
	if *onboardSpec != "" {
		defer closeQueueClient(queueClient)
		return runOnboard(context.Background(), st, queueClient, *onboardSpec)
	}

	// Serve path (not -onboard): build the auth stack + in-process worker + HTTP
	// surface and serve until shutdown.
	return runServer(&cfg, st, searchClient, queueClient)
}

// runServer builds the site-user auth stack, the in-process ingest worker, and
// the HTTP surface, then serves until a termination signal. It owns the session
// store and worker lifecycles; the caller owns the pool, and serveWithWorker
// owns the queue client on the happy path.
func runServer(
	cfg *config.Config, st *store.Store, searchClient *search.Client, queueClient *queue.Client,
) error {
	// Site-user auth: build the enabled providers (OIDC discovery happens here,
	// bounded by a startup context) and the Redis session store, before the
	// worker so a discovery/Redis failure aborts startup cleanly.
	startCtx, startCancel := context.WithTimeout(context.Background(), oidcDiscoveryTimeout)
	defer startCancel()
	providers, err := buildAuthProviders(startCtx, cfg)
	if err != nil {
		closeQueueClient(queueClient)
		return fmt.Errorf("building auth providers: %w", err)
	}
	sessionStore, err := session.New(cfg.Store.RedisURL, cfg.Session.TTL, cookiesSecure(cfg))
	if err != nil {
		closeQueueClient(queueClient)
		return fmt.Errorf("building session store: %w", err)
	}
	defer closeSession(sessionStore)

	// The worker runs in-process alongside the HTTP server (single-binary ethos).
	// It builds a per-installation GitHub client per job via ingestRunner.
	runner := &ingestRunner{store: st, indexer: searchClient, github: cfg.GitHub}
	worker, err := queue.NewWorker(cfg.Store.RedisURL, workerConcurrency, runner)
	if err != nil {
		closeQueueClient(queueClient)
		return fmt.Errorf("building queue worker: %w", err)
	}
	if serr := worker.Start(); serr != nil {
		closeQueueClient(queueClient)
		return fmt.Errorf("starting queue worker: %w", serr)
	}

	if cfg.AuthDisabled() {
		slog.Warn("auth disabled (AUTH_PROVIDERS=none): the read API is open, " +
			"every request is served as the anonymous identity")
	}

	slog.Info("starting docz-api",
		"version", version,
		"commit", commit,
		"addr", cfg.HTTP.Addr,
		"auth_providers", cfg.Auth.Providers,
		"worker_concurrency", workerConcurrency,
	)

	router := newRouter([]namedChecker{
		{name: "postgres", check: st.Ping},
		{name: "meilisearch", check: searchClient.Health},
		{name: "redis", check: queueClient.Ping},
	})
	// The Prometheus scrape endpoint sits alongside the probes, outside the auth
	// gate (pull-based, typically scraped on the internal network).
	if cfg.Telemetry.MetricsEnabled {
		router.Handle("/metrics", telemetry.MetricsHandler())
	}
	// The /api/v1 gate is the session authn check (401 without a valid session)
	// composed over the authorize seam (which still grants all onboarded repos;
	// the future SpiceDB resolver plugs in here). Session runs first so authorize
	// resolves behind a real identity.
	authorizer := authorize.NewAllReposAuthorizer(st)
	gate := func(next http.Handler) http.Handler {
		return authn(cfg, sessionStore)(authorize.Middleware(authorizer)(next))
	}
	authHandler := authhttp.New(
		auth.NewRegistry(providers), sessionStore, st, []byte(cfg.Session.Secret.Reveal()),
	)
	// The read/search API plus the two session-gated auth endpoints share the gate.
	httpapi.NewHandlerWithSearch(st, searchClient).Mount(router, gate, authHandler.MountAPI)
	if !cfg.AuthDisabled() {
		// /auth/login and /auth/callback are public (state is their CSRF guard).
		// Under none-mode there is no provider to redirect to, so the routes are
		// left unmounted and 404 — the absent-route precedent from search.
		authHandler.MountPublic(router)
	}

	// The GitHub webhook receiver sits outside /api/v1 and the auth gate: it is
	// authenticated by its HMAC signature, not a session. It inherits the
	// router's RequestID + Recoverer middleware and enqueues onto the same queue.
	webhookHandler := webhook.New(
		[]byte(cfg.GitHub.WebhookSecret.Reveal()), st, queueClient, searchClient,
	)
	router.Post("/webhooks/github", webhookHandler.ServeHTTP)

	return serveWithWorker(cfg.HTTP.Addr, router, worker, queueClient)
}

// workerConcurrency bounds the number of parallel ingest jobs. Two balances
// homelab resource use with throughput: each job holds a pool connection and
// issues GitHub API calls.
const workerConcurrency = 2

// runOnboard seeds an installation and enqueues one ingest for spec
// (owner/name@installation_id), then returns. The manual trigger now goes
// through the queue — a running server's worker performs the ingest. Phase 5
// drives the same enqueue from webhooks.
func runOnboard(ctx context.Context, st *store.Store, enq queue.Enqueuer, spec string) error {
	owner, name, installationID, err := parseOnboardSpec(spec)
	if err != nil {
		return fmt.Errorf("parse -onboard: %w", err)
	}

	// Account login/type are placeholders here; the installation webhook
	// (Phase 5) overwrites them with the real values on the next event. The
	// installation row must exist before the worker reconciles (repos FK it).
	if uerr := st.UpsertInstallation(ctx, store.InstallationInput{
		ID:           installationID,
		AccountLogin: owner,
		AccountType:  "Organization",
	}); uerr != nil {
		return fmt.Errorf("seed installation: %w", uerr)
	}

	if eerr := enq.EnqueueIngest(ctx, &queue.IngestJob{
		InstallationID: installationID,
		Owner:          owner,
		Name:           name,
		Reason:         "onboard",
	}); eerr != nil {
		return fmt.Errorf("enqueue onboard ingest: %w", eerr)
	}

	slog.Info("onboard enqueued; a running worker will ingest shortly",
		"repo", owner+"/"+name, "installation_id", installationID)
	return nil
}

// parseOnboardSpec splits "owner/name@installation_id" into its parts.
func parseOnboardSpec(spec string) (owner, name string, installationID int64, err error) {
	repoPart, idPart, ok := strings.Cut(spec, "@")
	if !ok {
		return "", "", 0, fmt.Errorf("expected owner/name@installation_id, got %q", spec)
	}
	owner, name, ok = strings.Cut(repoPart, "/")
	if !ok || owner == "" || name == "" {
		return "", "", 0, fmt.Errorf("expected owner/name, got %q", repoPart)
	}
	installationID, err = strconv.ParseInt(idPart, 10, 64)
	if err != nil {
		return "", "", 0, fmt.Errorf("invalid installation id %q: %w", idPart, err)
	}
	if installationID <= 0 {
		return "", "", 0, fmt.Errorf("installation id must be positive, got %d", installationID)
	}
	return owner, name, installationID, nil
}

// newLogger builds the slog handler selected by the log config. Config
// validation already constrains level/format, so the error paths are a
// defensive backstop.
func newLogger(cfg config.LogConfig) (*slog.Logger, error) {
	var level slog.Level
	switch cfg.Level {
	case "debug":
		level = slog.LevelDebug
	case "info":
		level = slog.LevelInfo
	case "warn":
		level = slog.LevelWarn
	case "error":
		level = slog.LevelError
	default:
		return nil, fmt.Errorf("invalid log level %q", cfg.Level)
	}

	opts := &slog.HandlerOptions{Level: level}
	switch cfg.Format {
	case "json":
		return slog.New(slog.NewJSONHandler(os.Stdout, opts)), nil
	case "text":
		return slog.New(slog.NewTextHandler(os.Stdout, opts)), nil
	default:
		return nil, fmt.Errorf("invalid log format %q", cfg.Format)
	}
}

// newRouter builds the base chi router with the liveness and readiness probes.
// The /api/v1 read routes are mounted onto it by the caller so they sit behind
// the authorize middleware while the probes stay open.
func newRouter(checkers []namedChecker) chi.Router {
	r := chi.NewRouter()
	r.Use(middleware.RequestID)
	// RequestLogger + Instrument sit outside Recoverer so a recovered panic is
	// still logged and metered as a 500; probe paths are skipped inside them.
	r.Use(telemetry.RequestLogger(slog.Default()))
	r.Use(telemetry.Instrument())
	r.Use(middleware.Recoverer)
	r.Get("/healthz", handleHealthz)
	r.Get("/readyz", handleReadyz(checkers))
	// The OpenAPI contract is public (outside the /api/v1 auth gate) so consumers
	// like the docz-site can fetch the machine contract without a session. There
	// is no /docs UI page (IMPL-0002 OQ-3d); this raw spec is the only surface.
	r.Get("/openapi.yaml", handleOpenAPISpec)
	return r
}

// handleOpenAPISpec serves the embedded OpenAPI contract (api.Spec) verbatim,
// so the served bytes are provably identical to api/openapi.yaml and to the copy
// the contract test validates. No filesystem dependency — it works in the
// distroless image.
func handleOpenAPISpec(w http.ResponseWriter, _ *http.Request) {
	w.Header().Set("Content-Type", "application/yaml")
	w.WriteHeader(http.StatusOK)
	if _, err := w.Write(api.Spec); err != nil {
		slog.Debug("openapi spec write failed", "err", err)
	}
}

// handleHealthz is the liveness probe: it reports that the process is up,
// independent of downstream dependencies (readiness covers those).
func handleHealthz(w http.ResponseWriter, _ *http.Request) {
	writeJSON(w, http.StatusOK, `{"status":"ok"}`)
}

// handleReadyz is the readiness probe: it checks every named dependency and
// reports each one's status (e.g. {"meilisearch":"ok","postgres":"ok"}). It
// returns 503 if any check fails, 200 otherwise, so a rollout or dependency
// outage gates traffic and the body names the offender.
func handleReadyz(checkers []namedChecker) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		ctx, cancel := context.WithTimeout(r.Context(), readyzTimeout)
		defer cancel()

		status := make(map[string]string, len(checkers))
		healthy := true
		for _, c := range checkers {
			if err := c.check(ctx); err != nil {
				slog.Warn("readiness check failed", "dep", c.name, "err", err)
				status[c.name] = "unavailable"
				healthy = false
				continue
			}
			status[c.name] = "ok"
		}

		code := http.StatusOK
		if !healthy {
			code = http.StatusServiceUnavailable
		}
		writeReadyz(w, code, status)
	}
}

// writeReadyz writes the readiness status map as JSON with the given code. Map
// keys marshal in sorted order, so the body is deterministic.
func writeReadyz(w http.ResponseWriter, code int, status map[string]string) {
	body, err := json.Marshal(status)
	if err != nil {
		// status is a map[string]string; marshaling cannot realistically fail.
		body = []byte(`{"status":"error"}`)
		code = http.StatusInternalServerError
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(code)
	if _, werr := w.Write(body); werr != nil {
		slog.Debug("readyz write failed", "err", werr)
	}
}

// writeJSON writes a status code and a JSON body, logging a failed write at
// debug (the client has already gone away; nothing else to do).
func writeJSON(w http.ResponseWriter, status int, body string) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	if _, err := w.Write([]byte(body)); err != nil {
		slog.Debug("response write failed", "status", status, "err", err)
	}
}

// serveWithWorker runs the HTTP server and the async ingest worker until an
// interrupt/terminate signal, then drains in order: HTTP first (so no new
// webhook/onboard enqueues arrive), then the worker (drain in-flight ingests),
// then the queue client. This ordering guarantees no ingest job is lost.
func serveWithWorker(addr string, handler http.Handler, w *queue.Worker, qc *queue.Client) error {
	srv := &http.Server{
		Addr:              addr,
		Handler:           handler,
		ReadHeaderTimeout: readHeaderTimeout,
	}

	ctx, stop := signal.NotifyContext(context.Background(), os.Interrupt, syscall.SIGTERM)
	defer stop()

	serveErr := make(chan error, 1)
	go func() {
		slog.Info("http server listening", "addr", addr)
		if err := srv.ListenAndServe(); err != nil && !errors.Is(err, http.ErrServerClosed) {
			serveErr <- err
		}
	}()

	select {
	case err := <-serveErr:
		w.Shutdown()
		closeQueueClient(qc)
		return fmt.Errorf("http server: %w", err)
	case <-ctx.Done():
		slog.Info("shutdown signal received, draining")
	}

	// 1) Stop accepting HTTP and drain in-flight requests (no new enqueues).
	shutdownCtx, cancel := context.WithTimeout(context.Background(), shutdownTimeout)
	defer cancel()
	if err := srv.Shutdown(shutdownCtx); err != nil {
		slog.Warn("http shutdown did not complete cleanly", "err", err)
	}
	// 2) Drain in-flight ingest jobs.
	slog.Info("draining ingest worker")
	w.Shutdown()
	// 3) Close the enqueue/redis client.
	closeQueueClient(qc)
	slog.Info("shutdown complete")
	return nil
}

// closeQueueClient closes the queue client, logging a close error at warn.
func closeQueueClient(qc *queue.Client) {
	if err := qc.Close(); err != nil {
		slog.Warn("close queue client", "err", err)
	}
}

// closeSession closes the session store, logging a close error at warn.
func closeSession(s *session.Store) {
	if err := s.Close(); err != nil {
		slog.Warn("close session store", "err", err)
	}
}

// shutdownTelemetry flushes any pending spans, bounded by shutdownTimeout. It is
// a no-op when tracing is disabled.
func shutdownTelemetry(shutdown func(context.Context) error) {
	ctx, cancel := context.WithTimeout(context.Background(), shutdownTimeout)
	defer cancel()
	if err := shutdown(ctx); err != nil {
		slog.Warn("telemetry shutdown", "err", err)
	}
}

// githubCheckTimeout bounds the startup App-credential check. It matches the
// OIDC discovery budget: both are one startup round trip to an identity
// provider, and neither should stall a deploy for long.
const githubCheckTimeout = 15 * time.Second

// checkGitHubCredentials authenticates as the App at startup so a bad app id or
// private key fails the deploy loudly instead of surfacing later as an ingest
// that cannot succeed (INV-0007 F5).
//
// Only a rejection by GitHub is fatal. Being unable to reach GitHub is
// transient and logged as a warning: the read API serves fine from Postgres and
// Meilisearch while GitHub is unavailable, so a GitHub outage during a pod
// restart must not turn into a crash loop.
func checkGitHubCredentials(ctx context.Context, cfg config.GitHubConfig) error {
	ctx, cancel := context.WithTimeout(ctx, githubCheckTimeout)
	defer cancel()

	app, err := githubapp.SelfCheck(ctx, cfg.AppID, []byte(cfg.PrivateKey.Reveal()), cfg.APIBase)
	switch {
	case err == nil:
		slog.Info("github app authenticated",
			"app_id", app.ID, "app_slug", app.Slug, "app_name", app.Name)
		return nil
	case errors.Is(err, githubapp.ErrCredentialsRejected):
		return fmt.Errorf("checking github app credentials: %w", err)
	default:
		slog.Warn("could not verify github app credentials at startup; continuing",
			"err", err)
		return nil
	}
}

// authn returns the authentication middleware for the /api/v1 gate: the real
// session check normally, or the anonymous injector under AUTH_PROVIDERS=none.
// Both inject a session.Session, so everything downstream — authorize, the
// handlers, /api/v1/auth/session — is identical in either mode.
func authn(cfg *config.Config, sessionStore *session.Store) func(http.Handler) http.Handler {
	if cfg.AuthDisabled() {
		return session.AnonymousMiddleware()
	}
	return session.Middleware(sessionStore)
}